Collision logic refactor.
